+++ src/NumSharp.Core/Backends/Default/Math/Default.ClipNDArray.cs
@@ -79,7 +79,7 @@ namespace NumSharp.Backends
         /// <summary>
         /// Fast path for contiguous arrays - uses IL kernel with SIMD support.
         /// </summary>
-        private unsafe NDArray ClipNDArrayContiguous(NDArray @out, NDArray min, NDArray max, int len)
+        private unsafe NDArray ClipNDArrayContiguous(NDArray @out, NDArray min, NDArray max, long len)
         {
             if (!(min is null) && !(max is null))
             {
@@ -212,7 +212,7 @@ namespace NumSharp.Backends
         /// <summary>
         /// General path for non-contiguous/broadcast arrays - uses GetAtIndex for element access.
         /// </summary>
-        private unsafe NDArray ClipNDArrayGeneral(NDArray @out, NDArray min, NDArray max, int len)
+        private unsafe NDArray ClipNDArrayGeneral(NDArray @out, NDArray min, NDArray max, long len)
         {
             if (!(min is null) && !(max is null))
             {
@@ -341,7 +341,7 @@ namespace NumSharp.Backends
 
         #region General Path Core Methods
 
-        private static unsafe void ClipNDArrayGeneralCore<T>(NDArray @out, NDArray min, NDArray max, int len)
+        private static unsafe void ClipNDArrayGeneralCore<T>(NDArray @out, NDArray min, NDArray max, long len)
             where T : unmanaged, IComparable<T>
         {
             // Use specialized implementations for float/double to handle NaN correctly
@@ -357,9 +357,9 @@ namespace NumSharp.Backends
             }
 
             var outAddr = (T*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var minVal = Converts.ChangeType<T>(min.GetAtIndex(i));
                 var maxVal = Converts.ChangeType<T>(max.GetAtIndex(i));
@@ -373,7 +373,7 @@ namespace NumSharp.Backends
             }
         }
 
-        private static unsafe void ClipNDArrayMinGeneralCore<T>(NDArray @out, NDArray min, int len)
+        private static unsafe void ClipNDArrayMinGeneralCore<T>(NDArray @out, NDArray min, long len)
             where T : unmanaged, IComparable<T>
         {
             // Use specialized implementations for float/double to handle NaN correctly
@@ -389,9 +389,9 @@ namespace NumSharp.Backends
             }
 
             var outAddr = (T*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var minVal = Converts.ChangeType<T>(min.GetAtIndex(i));
 
@@ -400,7 +400,7 @@ namespace NumSharp.Backends
             }
         }
 
-        private static unsafe void ClipNDArrayMaxGeneralCore<T>(NDArray @out, NDArray max, int len)
+        private static unsafe void ClipNDArrayMaxGeneralCore<T>(NDArray @out, NDArray max, long len)
             where T : unmanaged, IComparable<T>
         {
             // Use specialized implementations for float/double to handle NaN correctly
@@ -416,9 +416,9 @@ namespace NumSharp.Backends
             }
 
             var outAddr = (T*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var maxVal = Converts.ChangeType<T>(max.GetAtIndex(i));
 
@@ -431,12 +431,12 @@ namespace NumSharp.Backends
 
         // These use Math.Max/Min which properly propagate NaN per IEEE semantics
 
-        private static unsafe void ClipNDArrayGeneralCoreFloat(NDArray @out, NDArray min, NDArray max, int len)
+        private static unsafe void ClipNDArrayGeneralCoreFloat(NDArray @out, NDArray min, NDArray max, long len)
         {
             var outAddr = (float*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var minVal = Converts.ToSingle(min.GetAtIndex(i));
                 var maxVal = Converts.ToSingle(max.GetAtIndex(i));
@@ -444,12 +444,12 @@ namespace NumSharp.Backends
             }
         }
 
-        private static unsafe void ClipNDArrayGeneralCoreDouble(NDArray @out, NDArray min, NDArray max, int len)
+        private static unsafe void ClipNDArrayGeneralCoreDouble(NDArray @out, NDArray min, NDArray max, long len)
         {
             var outAddr = (double*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var minVal = Converts.ToDouble(min.GetAtIndex(i));
                 var maxVal = Converts.ToDouble(max.GetAtIndex(i));
@@ -457,48 +457,48 @@ namespace NumSharp.Backends
             }
         }
 
-        private static unsafe void ClipNDArrayMinGeneralCoreFloat(NDArray @out, NDArray min, int len)
+        private static unsafe void ClipNDArrayMinGeneralCoreFloat(NDArray @out, NDArray min, long len)
         {
             var outAddr = (float*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var minVal = Converts.ToSingle(min.GetAtIndex(i));
                 outAddr[outOffset] = Math.Max(val, minVal);
             }
         }
 
-        private static unsafe void ClipNDArrayMinGeneralCoreDouble(NDArray @out, NDArray min, int len)
+        private static unsafe void ClipNDArrayMinGeneralCoreDouble(NDArray @out, NDArray min, long len)
         {
             var outAddr = (double*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var minVal = Converts.ToDouble(min.GetAtIndex(i));
                 outAddr[outOffset] = Math.Max(val, minVal);
             }
         }
 
-        private static unsafe void ClipNDArrayMaxGeneralCoreFloat(NDArray @out, NDArray max, int len)
+        private static unsafe void ClipNDArrayMaxGeneralCoreFloat(NDArray @out, NDArray max, long len)
         {
             var outAddr = (float*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var maxVal = Converts.ToSingle(max.GetAtIndex(i));
                 outAddr[outOffset] = Math.Min(val, maxVal);
             }
         }
 
-        private static unsafe void ClipNDArrayMaxGeneralCoreDouble(NDArray @out, NDArray max, int len)
+        private static unsafe void ClipNDArrayMaxGeneralCoreDouble(NDArray @out, NDArray max, long len)
         {
             var outAddr = (double*)@out.Address;
-            for (int i = 0; i < len; i++)
+            for (long i = 0; i < len; i++)
             {
-                int outOffset = @out.Shape.TransformOffset(i);
+                long outOffset = @out.Shape.TransformOffset(i);
                 var val = outAddr[outOffset];
                 var maxVal = Converts.ToDouble(max.GetAtIndex(i));
                 outAddr[outOffset] = Math.Min(val, maxVal);
@@ -511,9 +511,9 @@ namespace NumSharp.Backends
 
         #region Scalar Fallbacks for Non-SIMD Types (Decimal, Char) - Array Bounds
 
-        private static unsafe void ClipArrayBoundsDecimal(decimal* output, decimal* minArr, decimal* maxArr, int size)
+        private static unsafe void ClipArrayBoundsDecimal(decimal* output, decimal* minArr, decimal* maxArr, long size)
         {
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
             {
                 var val = output[i];
                 if (val < minArr[i]) val = minArr[i];
@@ -522,21 +522,21 @@ namespace NumSharp.Backends
             }
         }
 
-        private static unsafe void ClipArrayMinDecimal(decimal* output, decimal* minArr, int size)
+        private static unsafe void ClipArrayMinDecimal(decimal* output, decimal* minArr, long size)
         {
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
                 if (output[i] < minArr[i]) output[i] = minArr[i];
         }
 
-        private static unsafe void ClipArrayMaxDecimal(decimal* output, decimal* maxArr, int size)
+        private static unsafe void ClipArrayMaxDecimal(decimal* output, decimal* maxArr, long size)
         {
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
                 if (output[i] > maxArr[i]) output[i] = maxArr[i];
         }
 
-        private static unsafe void ClipArrayBoundsChar(char* output, char* minArr, char* maxArr, int size)
+        private static unsafe void ClipArrayBoundsChar(char* output, char* minArr, char* maxArr, long size)
         {
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
             {
                 var val = output[i];
                 if (val < minArr[i]) val = minArr[i];
@@ -545,15 +545,15 @@ namespace NumSharp.Backends
             }
         }
 
-        private static unsafe void ClipArrayMinChar(char* output, char* minArr, int size)
+        private static unsafe void ClipArrayMinChar(char* output, char* minArr, long size)
         {
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
                 if (output[i] < minArr[i]) output[i] = minArr[i];
         }
 
-        private static unsafe void ClipArrayMaxChar(char* output, char* maxArr, int size)
+        private static unsafe void ClipArrayMaxChar(char* output, char* maxArr, long size)
         {
-            for (int i = 0; i < size; i++)
+            for (long i = 0; i < size; i++)
                 if (output[i] > maxArr[i]) output[i] = maxArr[i];
         }
