+++ src/NumSharp.Core/Backends/Default/Math/Reduction/Default.Reduction.Add.cs
@@ -40,7 +40,7 @@ namespace NumSharp.Backends
             var result = sum_elementwise_il(arr, typeCode);
             if (@out != null) { @out.SetAtIndex(result, 0); return @out; }
             var r = NDArray.Scalar(result);
-            if (keepdims) { var ks = new int[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
+            if (keepdims) { var ks = new long[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
             else if (!r.Shape.IsScalar && r.Shape.size == 1 && r.ndim == 1) r.Storage.Reshape(Shape.Scalar);
             return r;
         }
@@ -54,7 +54,7 @@ namespace NumSharp.Backends
             if (kernel == null)
                 throw new NotSupportedException($"Axis reduction kernel not available for {op}({inputType}) -> {outputType}.");
 
-            var outputDims = new int[arr.ndim - 1];
+            var outputDims = new long[arr.ndim - 1];
             for (int d = 0, od = 0; d < arr.ndim; d++) if (d != axis) outputDims[od++] = shape.dimensions[d];
 
             var outputShape = outputDims.Length > 0 ? new Shape(outputDims) : Shape.Scalar;
@@ -62,20 +62,20 @@ namespace NumSharp.Backends
             if (@out != null) { if (@out.Shape != outputShape) throw new IncorrectShapeException($"Output shape mismatch"); result = @out; }
             else result = new NDArray(outputType, outputShape, false);
 
-            int axisSize = shape.dimensions[axis];
-            int outputSize = result.size > 0 ? result.size : 1;
+            long axisSize = shape.dimensions[axis];
+            long outputSize = result.size > 0 ? result.size : 1;
             byte* inputAddr = (byte*)arr.Address + shape.offset * arr.dtypesize;
 
-            fixed (int* inputStrides = shape.strides)
-            fixed (int* inputDims = shape.dimensions)
-            fixed (int* outputStrides = result.Shape.strides)
+            fixed (long* inputStrides = shape.strides)
+            fixed (long* inputDims = shape.dimensions)
+            fixed (long* outputStrides = result.Shape.strides)
             {
                 kernel((void*)inputAddr, (void*)result.Address, inputStrides, inputDims, outputStrides, axis, axisSize, arr.ndim, outputSize);
             }
 
             if (keepdims)
             {
-                var ks = new int[arr.ndim];
+                var ks = new long[arr.ndim];
                 for (int d = 0, sd = 0; d < arr.ndim; d++) ks[d] = (d == axis) ? 1 : result.shape[sd++];
                 result.Storage.Reshape(new Shape(ks));
             }
@@ -112,7 +112,7 @@ namespace NumSharp.Backends
 
             if (keepdims)
             {
-                var ks = new int[arr.ndim];
+                var ks = new long[arr.ndim];
                 for (int d = 0, sd = 0; d < arr.ndim; d++)
                     ks[d] = (d == axis) ? 1 : resultShape[sd++];
                 result.Storage.Reshape(new Shape(ks));
@@ -128,7 +128,7 @@ namespace NumSharp.Backends
                 var defaultVal = (typeCode ?? arr.typecode).GetDefaultValue();
                 if (@out != null) { @out.SetAtIndex(defaultVal, 0); return @out; }
                 var r = NDArray.Scalar(defaultVal);
-                if (keepdims) { var ks = new int[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
+                if (keepdims) { var ks = new long[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
                 return r;
             }
             var axis = NormalizeAxis(axis_.Value, arr.ndim);
@@ -137,7 +137,7 @@ namespace NumSharp.Backends
             var result = np.zeros(new Shape(resultShape), outputType);
             if (keepdims)
             {
-                var ks = new int[arr.ndim];
+                var ks = new long[arr.ndim];
                 for (int d = 0, sd = 0; d < arr.ndim; d++) ks[d] = (d == axis) ? 1 : resultShape[sd++];
                 result.Storage.Reshape(new Shape(ks));
             }
@@ -149,7 +149,7 @@ namespace NumSharp.Backends
         {
             var r = typeCode.HasValue ? Cast(arr, typeCode.Value, true) : arr.Clone();
             if (@out != null) { @out.SetAtIndex(r.GetAtIndex(0), 0); return @out; }
-            if (keepdims) { var ks = new int[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
+            if (keepdims) { var ks = new long[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
             else if (!r.Shape.IsScalar && r.Shape.size == 1 && r.ndim == 1) r.Storage.Reshape(Shape.Scalar);
             return r;
         }
@@ -158,9 +158,9 @@ namespace NumSharp.Backends
         {
             if (@out != null) return null;
             var shape = arr.Shape;
-            int[] resultDims;
+            long[] resultDims;
             if (keepdims) { resultDims = (long[])shape.dimensions.Clone(); resultDims[axis] = 1; }
-            else { resultDims = new int[arr.ndim - 1]; for (int d = 0, rd = 0; d < arr.ndim; d++) if (d != axis) resultDims[rd++] = shape.dimensions[d]; }
+            else { resultDims = new long[arr.ndim - 1]; for (int d = 0, rd = 0; d < arr.ndim; d++) if (d != axis) resultDims[rd++] = shape.dimensions[d]; }
             if (resultDims.Length == 0)
             {
                 var v = arr.GetAtIndex(0);
@@ -168,8 +168,8 @@ namespace NumSharp.Backends
                 return NDArray.Scalar(v);
             }
             var result = new NDArray(outputType, new Shape(resultDims), false);
-            if (outputType == arr.GetTypeCode) for (int i = 0; i < result.size; i++) result.SetAtIndex(arr.GetAtIndex(i), i);
-            else for (int i = 0; i < result.size; i++) result.SetAtIndex(Converts.ChangeType(arr.GetAtIndex(i), outputType), i);
+            if (outputType == arr.GetTypeCode) for (long i = 0; i < result.size; i++) result.SetAtIndex(arr.GetAtIndex(i), i);
+            else for (long i = 0; i < result.size; i++) result.SetAtIndex(Converts.ChangeType(arr.GetAtIndex(i), outputType), i);
             return result;
         }
